chore: move cled_controller.h into fl/channels/ + harden CI workflows
## Source move
Relocates `src/fl/cled_controller.h` to `src/fl/channels/cled_controller.h`
so the LED controller base class lives next to the channels/buses that
wrap it (and the `ChannelOptions` it already depends on). The legacy
shim `src/cled_controller.h` and 2 direct test consumers are updated to
the new path; the user-facing `using CLEDController = fl::CLEDController`
re-export is unchanged — no API change.

## CI workflow normalization
Removes the `trusted/pr-code` overlay dance from 10 reusable workflows:
- build_template.yml (called by ~30 board builds)
- template_unit_test.yml, template_example_test.yml
- build_wasm.yml, build_wasm_compilers.yml, build_linux.yml
- qemu_docker_template.yml, avr8js_docker_template.yml
- build_template_custom_board.yml, build_template_binary_size.yml

The 4-step pattern (checkout master to `trusted/`, checkout PR head to
`pr-code/`, `cp -r pr-code/src/* trusted/src/`, override
`GITHUB_WORKSPACE`) dates to the `pull_request_target` era when the
project ran with secrets in scope and wanted master's CI scripts to run
against PR-supplied source. That trigger is no longer used (only
`pull_request`, with read-only tokens for forks); external PRs are
gated through other mechanisms.

Two concrete problems with the old pattern:
1. `cp -r` is an overlay merge that does not propagate file deletions —
   any PR that moves or removes a file produced duplicate-symbol
   failures in CI even when local builds passed cleanly.
2. The `trusted/` workspace ran a "Configure Git/LFS auth" step that
   wrote `http.https://github.com/.extraheader` to global git config,
   which leaked GITHUB_TOKEN to subsequent steps including ./compile
   running PR code (CodeRabbit, comment 2).

Each affected workflow is replaced with a standard single-checkout
pattern. Net: −493/+40 lines across the 10 files.

## Hardening (CodeRabbit findings)
- All 10 normalized checkouts now set `persist-credentials: false`
  (CodeRabbit, comment 1) so GITHUB_TOKEN is not persisted in
  `.git/config` for subsequent steps. Combined with `id-token: write`
  permission, this prevents PR-controlled code from minting OIDC tokens
  off the runner's identity.
- Drops the dead "Configure Git/LFS auth for ALL github.com traffic"
  step from build_template.yml — the repo does not use Git LFS
  (`.gitattributes` has zero `filter=lfs` directives), and the global
  config it set negated the persist-credentials protection.
