+++ .github/workflows/build_template_custom_board.yml
@@ -17,48 +17,10 @@ jobs:
   build:
     runs-on: ubuntu-latest
     steps:
-      - name: Checkout trusted base branch code
+      - name: Checkout
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.event.repository.default_branch || 'master' }}
-          path: trusted
-
-      - name: Checkout PR code
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.pull_request.head.sha || github.sha }}
-          path: pr-code
-
-      - name: Merge safe source files from PR
-        run: |
-          # Copy ONLY safe source files from PR, never executable scripts
-          if [ -d "pr-code/src" ]; then
-            cp -r pr-code/src/* trusted/src/ 2>/dev/null || true
-          fi
-          if [ -d "pr-code/examples" ]; then
-            cp -r pr-code/examples/* trusted/examples/ 2>/dev/null || true
-          fi
-          if [ -d "pr-code/ci/boards" ]; then
-            cp -r pr-code/ci/boards/* trusted/ci/boards/ 2>/dev/null || true
-          fi
-          if [ -f "pr-code/platformio.ini" ]; then
-            cp pr-code/platformio.ini trusted/
-          fi
-          if [ -f "pr-code/library.json" ]; then
-            cp pr-code/library.json trusted/
-          fi
-          if [ -f "pr-code/library.properties" ]; then
-            cp pr-code/library.properties trusted/
-          fi
-          if [ -f "pr-code/CMakeLists.txt" ]; then
-            cp pr-code/CMakeLists.txt trusted/
-          fi
-          # Change working directory to trusted for all subsequent steps
-          cd trusted
-        shell: bash
-
-      - name: Set working directory to trusted
-        run: echo "GITHUB_WORKSPACE=$GITHUB_WORKSPACE/trusted" >> $GITHUB_ENV
+          persist-credentials: false
 
       - name: Install Python
         uses: actions/setup-python@v5
@@ -67,19 +29,16 @@ jobs:
 
       # Use the fastest build to instantiate the ~/.platformio/boards folder
       - name: Pre-build to ensure ~/.platformio/boards folder is created
-        working-directory: trusted
         run: ./compile uno --examples Blink
         shell: bash
 
       # Copy the custom board JSON file to the PlatformIO boards folder
       - name: Copy ci/boards/*.json to ~/.platformio/boards
-        working-directory: trusted
         run: |
           mkdir -p ~/.platformio/boards
           cp ci/boards/*.json ~/.platformio/boards
 
       - name: Build FastLED examples
-        working-directory: trusted
         run: |
           set -o pipefail
           ./compile --no-interactive --max-failures 3 ${{ inputs.args }} | tee build.log
@@ -87,7 +46,6 @@ jobs:
 
       - name: Build summary and failure logs
         if: always()
-        working-directory: trusted
         run: |
           SUMMARY_FILE="build-summary.txt"
           {
@@ -99,7 +57,6 @@ jobs:
         shell: bash
 
       - name: Print out build_info files
-        working-directory: trusted
         run: |
           build_info_files=$(find . -name 'build_info*.json')
           if [ -z "$build_info_files" ]; then
@@ -123,7 +80,6 @@ jobs:
         if: always()
 
       - name: CPP Check
-        working-directory: trusted
         run: |
           # take the input.args and parse out the first element, which will be the board name
           # do this to input.args
@@ -143,22 +99,21 @@ jobs:
         uses: actions/upload-artifact@v4
         with:
           name: build-artifacts-${{ steps.generate_id.outputs.timestamp }}-${{ github.sha }}-${{ steps.generate_id.outputs.random_hex }}
-          path: trusted/.build/
+          path: .build/
           include-hidden-files: true
-
 
       - name: Upload build log
         uses: actions/upload-artifact@v4
         with:
           name: build-log-${{ steps.generate_id.outputs.timestamp }}-${{ github.sha }}-${{ steps.generate_id.outputs.random_hex }}
-          path: trusted/build.log
+          path: build.log
 
       - name: Upload build summary
         if: always()
         uses: actions/upload-artifact@v4
         with:
           name: build-summary-${{ steps.generate_id.outputs.timestamp }}-${{ github.sha }}-${{ steps.generate_id.outputs.random_hex }}
-          path: trusted/build-summary.txt
+          path: build-summary.txt
           if-no-files-found: ignore
 
       - name: Check build status
