+++ .github/workflows/template_unit_test.yml
@@ -22,61 +22,10 @@ jobs:
   build:
     runs-on: ${{ inputs.os }}
     steps:
-      - name: Checkout trusted base branch code
+      - name: Checkout
         uses: actions/checkout@v4
         with:
-          ref: ${{ github.event.repository.default_branch || 'master' }}
-          path: trusted
-
-      - name: Checkout PR code
-        uses: actions/checkout@v4
-        with:
-          ref: ${{ github.event.pull_request.head.sha || github.sha }}
-          path: pr-code
-
-      - name: Merge safe source files from PR
-        run: |
-          # Copy ONLY safe source files from PR, never executable scripts
-          if [ -d "pr-code/src" ]; then
-            cp -r pr-code/src/* trusted/src/ 2>/dev/null || true
-          fi
-          if [ -d "pr-code/tests" ]; then
-            cp -r pr-code/tests/* trusted/tests/ 2>/dev/null || true
-          fi
-          if [ -d "pr-code/examples" ]; then
-            cp -r pr-code/examples/* trusted/examples/ 2>/dev/null || true
-          fi
-          if [ -f "pr-code/platformio.ini" ]; then
-            cp pr-code/platformio.ini trusted/
-          fi
-          if [ -f "pr-code/library.json" ]; then
-            cp pr-code/library.json trusted/
-          fi
-          if [ -f "pr-code/library.properties" ]; then
-            cp pr-code/library.properties trusted/
-          fi
-          if [ -f "pr-code/CMakeLists.txt" ]; then
-            cp pr-code/CMakeLists.txt trusted/
-          fi
-          if [ -f "pr-code/test.py" ]; then
-            cp pr-code/test.py trusted/
-          fi
-          if [ "${{ github.event.pull_request.head.repo.full_name }}" = "${{ github.repository }}" ]; then
-            # Same-repository CI helper PRs need to exercise their own Meson setup code.
-            if [ -f "pr-code/ci/meson/build_config.py" ]; then
-              cp pr-code/ci/meson/build_config.py trusted/ci/meson/build_config.py
-            fi
-          fi
-          if [ -f "pr-code/test" ]; then
-            # DO NOT copy executable test script - use trusted version
-            echo "Skipping pr-code/test script - using trusted version"
-          fi
-          # Change working directory to trusted for all subsequent steps
-          cd trusted
-        shell: bash
-
-      - name: Set working directory to trusted
-        run: echo "GITHUB_WORKSPACE=$GITHUB_WORKSPACE/trusted" >> $GITHUB_ENV
+          persist-credentials: false
 
       - name: Install Python
         uses: actions/setup-python@v5
@@ -95,18 +44,15 @@ jobs:
         run: sudo apt-get install -y libunwind-dev
 
       - name: Install
-        working-directory: trusted
         run: ./install
         shell: bash
 
       - name: No namespace after headers
-        working-directory: trusted
         run: |
           uv run python ci/lint_cpp/no_using_namespace_fl_in_headers.py
 
       - name: Run All Unit Tests
         id: run_tests
-        working-directory: trusted
         run: |
           set -o pipefail
           ./test --clang --no-parallel --unit --verbose --debug --log-failures failures 2>&1 | tee test.log
@@ -114,7 +60,6 @@ jobs:
 
       - name: Test summary and failure logs
         if: always()
-        working-directory: trusted
         run: |
           echo "############################################"
           echo "Last 100 lines of test.log (most recent output)"
