+++ src/cled_controller.h
@@ -4,7 +4,7 @@
 /// @file cled_controller.h
 /// base definitions used by led controllers for writing out led data
 
-#include "fl/cled_controller.h"
+#include "fl/channels/cled_controller.h"
 
 // Backward compatibility: bring fl::CLEDController into global namespace
 using CLEDController = fl::CLEDController;
